完成drop table、insert和条件查询，修复编译过程中出现的warning

+++ src/execution/execution_manager.cpp
@@ -161,7 +161,7 @@ void QlManager::select_from(std::unique_ptr<AbstractExecutor> executorTreeRoot, 
     std::fstream outfile;
     outfile.open("output.txt", std::ios::out | std::ios::app);
     outfile << "|";
-    for(int i = 0; i < captions.size(); ++i) {
+    for(int i = 0; i < (int)captions.size(); ++i) {
         outfile << " " << captions[i] << " |";
     }
     outfile << "\n";
@@ -189,7 +189,7 @@ void QlManager::select_from(std::unique_ptr<AbstractExecutor> executorTreeRoot, 
         rec_printer.print_record(columns, context);
         // print record into file
         outfile << "|";
-        for(int i = 0; i < columns.size(); ++i) {
+        for(int i = 0; i < (int)columns.size(); ++i) {
             outfile << " " << columns[i] << " |";
         }
         outfile << "\n";
